Move ray and spark plugins/activities into CMD instead of core modules
