+++ go/cmd/worker/BUILD.bazel
@@ -12,6 +12,8 @@ go_library(
         "//go/base/env:go_default_library",
         "//go/base/zapfx:go_default_library",
         "//go/worker:go_default_library",
+        "//go/worker/activities/ray:go_default_library",
+        "//go/worker/activities/spark:go_default_library",
         "@org_uber_go_fx//:go_default_library",
     ],
 )
