+++ go/worker/activities/BUILD.bazel
@@ -9,8 +9,6 @@ go_library(
         "//go/worker/activities/cachedoutput:go_default_library",
         "//go/worker/activities/model:go_default_library",
         "//go/worker/activities/notification:go_default_library",
-        "//go/worker/activities/ray:go_default_library",
-        "//go/worker/activities/spark:go_default_library",
         "//go/worker/activities/storage:go_default_library",
         "//go/worker/activities/trigger:go_default_library",
         "@org_uber_go_fx//:go_default_library",
